--- a/lib/data/entity.js
+++ b/lib/data/entity.js
@@ -238,7 +238,7 @@ const streamEntityCsv = (inStream, properties) => {
 
   // System properties
   header.push(...[ '__createdAt', '__creatorId', '__creatorName', '__updates', '__updatedAt', '__version']);
-  props.push(...['createdAt', 'creatorId', 'aux.creator.displayName', 'aux.stats.updates', 'updatedAt', 'def.version'].map(e => e.split('.')));
+  props.push(...['createdAt', 'creatorId', 'aux.creator.displayName', 'updates', 'updatedAt', 'def.version'].map(e => e.split('.')));
 
   const entityStream = _entityTransformer(header, props);
 
@@ -282,7 +282,7 @@ const selectFields = (entity, properties, selectedProperties) => {
     createdAt: entity.createdAt,
     creatorId: entity.aux.creator.id.toString(),
     creatorName: entity.aux.creator.displayName,
-    updates: entity.aux.stats.updates,
+    updates: entity.updates, // 'updates' is a property predating 'version' and will always equal version - 1
     updatedAt: entity.updatedAt,
     version: entity.def.version,
     conflict: entity.conflict