--- a/lib/data/entity.js
+++ b/lib/data/entity.js
@@ -282,7 +282,7 @@ const selectFields = (entity, properties, selectedProperties) => {
     createdAt: entity.createdAt,
     creatorId: entity.aux.creator.id.toString(),
     creatorName: entity.aux.creator.displayName,
-    updates: entity.updates, // 'updates' is a property predating 'version' and will always equal version - 1
+    updates: entity.updates,
     updatedAt: entity.updatedAt,
     version: entity.def.version,
     conflict: entity.conflict