--- a/lib/model/frames/entity.js
+++ b/lib/model/frames/entity.js
@@ -33,6 +33,14 @@ class Entity extends Frame.define(
 ) {
   get def() { return this.aux.def; }
 
+  // Note: we introduced this 'updates' count to entities in datasets (via odata and csv export)
+  // before we exposed a similar idea was exposed through the 'version' property.
+  // This version property increments with each entity update so it is safe to compute 'updates'
+  // from version instead of calculating it in less efficient ways.
+  // Note that 'updates' is used in too many places in frontend code and exposed in the odata/csv
+  // exports to remove it from the system properties.
+  get updates() { return this.aux.def.version - 1; }
+
   static fromParseEntityData(entityData, options = { create: true, update: false }) {
     const { dataset } = entityData.system;
     const { data } = entityData;