--- a/lib/model/frames/entity.js
+++ b/lib/model/frames/entity.js
@@ -34,7 +34,7 @@ class Entity extends Frame.define(
   get def() { return this.aux.def; }
 
   // Note: we introduced this 'updates' count to entities in datasets (via odata and csv export)
-  // before we exposed a similar idea was exposed through the 'version' property.
+  // before we exposed a similar idea through the 'version' property.
   // This version property increments with each entity update so it is safe to compute 'updates'
   // from version instead of calculating it in less efficient ways.
   // Note that 'updates' is used in too many places in frontend code and exposed in the odata/csv
@@ -87,10 +87,6 @@ class Entity extends Frame.define(
 }
 
 Entity.Partial = class extends Entity {};
-
-Entity.Extended = class extends Frame.define(
-  'updates'
-) { };
 
 Entity.Def = Frame.define(
   table('entity_defs', 'def'),