--- a/lib/model/query/entities.js
+++ b/lib/model/query/entities.js
@@ -744,16 +744,12 @@ resolveConflict.audit = (entity, dataset) => (log) => log('entity.update.resolve
 ////////////////////////////////////////////////////////////////////////////////
 // SERVING ENTITIES
 
-const _exportUnjoiner = unjoiner(Entity, Entity.Def, Entity.Extended.into('stats'), Actor.alias('actors', 'creator'));
+const _exportUnjoiner = unjoiner(Entity, Entity.Def, Actor.alias('actors', 'creator'));
 
 const streamForExport = (datasetId, options = QueryOptions.none) => ({ stream }) =>
   stream(sql`
 SELECT ${_exportUnjoiner.fields} FROM entity_defs
 INNER JOIN entities ON entities.id = entity_defs."entityId"
-INNER JOIN
-  (
-    SELECT "entityId", (COUNT(id) - 1) AS "updates" FROM entity_defs GROUP BY "entityId"
-  ) stats ON stats."entityId"=entity_defs."entityId"
 LEFT JOIN actors ON entities."creatorId"=actors.id
 ${options.skiptoken ? sql`
   INNER JOIN